Move FitbenchmarkFittingProblem to parsing/parse_fitbenchmark_data.py
